backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dynaconf>=3.2.12",
    "fastapi>=0.135.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "strands-agents>=1.29.0",
    "uuid-utils>=0.14.1",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.0"]
openai = [
    "strands-agents[openai]>=1.26.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
